backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
supabase
python-dotenv
scikit-learn
pandas
numpy
joblib
